Add files via upload
Energy Consumption Prediction Web App
This project is a web application built with Flask that predicts energy consumption based on the time of day and day of the week. The app uses a Linear Regression model trained on historical energy usage data to provide users with predictions and recommendations on when to optimize appliance usage for energy efficiency. The user-friendly interface allows users to input data (hour of the day and day of the week) and receive instant feedback on predicted energy consumption along with practical suggestions for saving energy.

Key Features:

Machine learning model using Linear Regression for energy consumption prediction.
User-friendly Flask web app with interactive form for inputting data.
Recommendations on optimal appliance usage to reduce energy consumption.
Built with Python, Flask, Pandas, and Scikit-learn.